BUG修改

+++ Personnel_MS/MS_client/src/components/pages/attendance_Mgmt/overtime_management/edit_overtime.vue
@@ -130,6 +130,7 @@
 					Authorization:`Bearer `+localStorage.getItem('access_token'),
 				},
 				fileList: [],
+				triRemoveFlag: true,
 				fileFlag: '',
 				formdata1: {
 				},
@@ -210,31 +211,34 @@
 				console.log("选中的fileList", fileList); 
 			},
 			handleRemove(file, fileList) {
-				// 移除文件
-				console.log("移除的file", file);
-				let index = this.fileList.indexOf(file);
-				fileList.splice(index, 0, file);
-				this.$confirm("此操作将永久删除, 是否继续?", "提示", {
-					confirmButtonText: "确定",
-					cancelButtonText: "取消",
-					type: "warning"
-				}).then(() => {
-					this.$axios.delete("/iem_hrm/file/deleteFile/" + file.fileId)
-						.then(res => {
-						let result = res.data.retMsg;
-						if ("操作成功" == result) {
-							this.$message({ type: "success", message: result });
-							fileList.splice(index, 1);
-						} else {
-							this.$message({ type: "error", message: result });
-						}
-						}).catch(e => {
-							console.log(e);
-							this.$message({ type: "error", message: e.retMsg });
+				if(this.triRemoveFlag) {
+					// 移除文件
+					console.log("移除的file", file);
+					let index = this.fileList.indexOf(file);
+					fileList.splice(index, 0, file);
+					this.$confirm("此操作将永久删除, 是否继续?", "提示", {
+						confirmButtonText: "确定",
+						cancelButtonText: "取消",
+						type: "warning"
+					}).then(() => {
+						this.$axios.delete("/iem_hrm/file/deleteFile/" + file.fileId)
+							.then(res => {
+							let result = res.data.retMsg;
+							if ("操作成功" == result) {
+								this.$message({ type: "success", message: result });
+								fileList.splice(index, 1);
+							} else {
+								this.$message({ type: "error", message: result });
+							}
+							}).catch(e => {
+								console.log(e);
+								this.$message({ type: "error", message: e.retMsg });
+						});
+					}).catch(() => {
+						this.$message({ type: "info", message: "已取消删除" });
 					});
-				}).catch(() => {
-					this.$message({ type: "info", message: "已取消删除" });
-				});
+				}
+				
 			},
 			handlePreview(file) {
 				// 点击已上传的文件链接时
@@ -267,8 +271,11 @@
 //		        console.log('上传文件只能是 xls、xlsx、doc、docx 格式!')
 //		      }
 		      if (!isLt2M) {
-		      	this.$message({ message: '上传文件大小不能超过 10MB!', type: 'error' });
-		      }
+				  this.$message({ message: '上传文件大小不能超过 10MB!', type: 'error' });
+				  this.triRemoveFlag = false;
+		      } else {
+				  this.triRemoveFlag = true;
+			  }
 		      return  isLt2M	//extension || extension2 || extension3 || extension4 &&
 		    },
 	      	save(formName) {
